+++ frontend/src/components/CalendarContainer.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Calendar } from '../hooks/useCalendar';
+import { Calendar } from '../hooks/calendar/useCalendar';
 interface Props {}
 
 export const CalendarContainer = (props: Props) => {
